Rename module to clients.py.
